% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize Regression Variables}
\usage{
visualize(model, measure = "P", signif = 0.05, sorted = TRUE)
}
\arguments{
\item{model}{an lm() object}

\item{measure}{use P-values (P) or coefficients (C) to construct plot}

\item{signif}{significance level for P-values}

\item{sorted}{whether the variables should be ordered by significance}
}
\value{
a ggplot2 object
}
\description{
Function to visually display the significance of each predictor
variable in a regression. The function takes an lm() object as input and
outputs a ggplot2-labeled figure.
}
\examples{
X <- runif(500,1,5)
Y <- runif(500,6,10)
Z <- runif(500,-5,0)

W <- rnorm(500,2*Y-X,5)

model <- lm(W~X+Y+Z)

visualize(model)
visualize(model, signif = 0.1)
visualize(model, measure = "C")
visualize(model, measure = "C", sorted = F)

}
