% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize Regression Variables}
\usage{
visualize(model, measure = "P")
}
\arguments{
\item{model}{an lm() object}

\item{measure}{use P-values (P), coefficients (C), or percent change (T) to construct plot}
}
\value{
a ggplot2 object
}
\description{
Function to visually display the significance of each predictor variable in
a regression. The function takes an lm() object as input and outputs a ggplot2-labeled
figure.
}
\examples{
X <- runif(500,1,5)
Y <- runif(500,6,10)

W <- rnorm(500,X+Y,5)

model <- lm(W~X+Y)

visualize(model)

}
